Style the SearchBar component
